@@ -1,15 +1,58 @@
+import 'package:flash_chat/screens/registration_screen.dart';
 import 'package:flutter/material.dart';
+import 'login_screen.dart';
+import 'package:animated_text_kit/animated_text_kit.dart';
+import 'package:flash_chat/RoundedButton.dart';
 
 class WelcomeScreen extends StatefulWidget {
+  static const String id = 'welcome_screen';
   @override
   _WelcomeScreenState createState() => _WelcomeScreenState();
 }
 
-class _WelcomeScreenState extends State<WelcomeScreen> {
+class _WelcomeScreenState extends State<WelcomeScreen> with SingleTickerProviderStateMixin{
+
+  late AnimationController controller;
+  late Animation animation;
+
+
+  @override
+  void initState() {
+    super.initState();
+
+    controller = AnimationController(
+      duration: Duration(seconds: 1),
+        vsync: this,
+      //upperbound means that by this much the ticker value will be increased
+    );
+
+    animation = ColorTween(begin: Color(0xffd0e8ef), end: Colors.white).animate(controller);
+    //This is going to proceed our animation forwards
+    controller.forward();
+    //In order to see what our controller is doing we can add a listener
+    //If we want our animation to go backwards then wee will use controller.reverse()
+    //syntax- controller.reverse(from: 1.0)
+    controller.addListener(() {
+      //setstate id called so that after every tick the UI is updated
+      setState(() {
+
+      });
+      print(animation.value);
+    });
+
+
+  }
+  //This animation will go on forever and will consume a lot of resources, so tap into the dispose method to ensure that when the screen is destroyed our controller is destroyed too.
+  @override
+  void dispose() {
+    super.dispose();
+    controller.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.white,
+      backgroundColor: animation.value,
       body: Padding(
         padding: EdgeInsets.symmetric(horizontal: 24.0),
         child: Column(
@@ -18,15 +61,24 @@ class _WelcomeScreenState extends State<WelcomeScreen> {
           children: <Widget>[
             Row(
               children: <Widget>[
-                Container(
-                  child: Image.asset('images/logo.png'),
-                  height: 60.0,
+                Hero(
+                  tag: 'logo',
+                  child: Container(
+                    margin: EdgeInsets.all(10.0),
+                    child: Image.asset('images/logo.png'),
+                    height: 60.0,
+                  ),
                 ),
-                Text(
-                  'Flash Chat',
+                DefaultTextStyle(
                   style: TextStyle(
                     fontSize: 45.0,
                     fontWeight: FontWeight.w900,
+                    color: Colors.grey
+                  ),
+                  child: AnimatedTextKit(
+                    animatedTexts: [
+                      TyperAnimatedText('Flash Chat'),],
+                    totalRepeatCount: 1,
                   ),
                 ),
               ],
@@ -34,45 +86,28 @@ class _WelcomeScreenState extends State<WelcomeScreen> {
             SizedBox(
               height: 48.0,
             ),
-            Padding(
-              padding: EdgeInsets.symmetric(vertical: 16.0),
-              child: Material(
-                elevation: 5.0,
-                color: Colors.lightBlueAccent,
-                borderRadius: BorderRadius.circular(30.0),
-                child: MaterialButton(
-                  onPressed: () {
-                    //Go to login screen.
-                  },
-                  minWidth: 200.0,
-                  height: 42.0,
-                  child: Text(
-                    'Log In',
-                  ),
-                ),
-              ),
+            RoundedButton(
+              text: 'Log In',
+              color: Colors.lightBlueAccent,
+              onPressed: (){
+                Navigator.pushNamed(
+                    context,
+                    LoginScreen.id);
+              }
             ),
-            Padding(
-              padding: EdgeInsets.symmetric(vertical: 16.0),
-              child: Material(
+            RoundedButton(
+                text: 'Register',
                 color: Colors.blueAccent,
-                borderRadius: BorderRadius.circular(30.0),
-                elevation: 5.0,
-                child: MaterialButton(
-                  onPressed: () {
-                    //Go to registration screen.
-                  },
-                  minWidth: 200.0,
-                  height: 42.0,
-                  child: Text(
-                    'Register',
-                  ),
-                ),
-              ),
-            ),
+                onPressed: (){
+                  Navigator.pushNamed(
+                      context,
+                      RegistrationScreen.id);
+                })
           ],
         ),
       ),
     );
   }
 }
+
+
